+++ packages/grid/src/features/standard-table/components/StandardTableCell.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import { KeyboardEventHandler, useCallback, useMemo } from "react";
 import { useGridCell } from "../../grid-cell/hooks/UseGridCell";
 import { useOnKeyDownContext } from "../context/OnKeyDownContext";
+import { useStickyPropsPerColumnContext } from "../context/StickyPropsPerColumnContext";
 import { useColumnIndexPerColumnIdContext } from "../features/column-index-per-column-id/ColumnIndexPerColumnIdContext";
 import { useCellBackgroundByColumnId } from "../hooks/UseCellBackground";
 import { useColumnConfigById } from "../hooks/UseColumnConfigById";
@@ -38,14 +39,13 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
     keyResolver,
     enableGridCell,
     gridCellOptions: gridCellOptionsForTable,
-    stickyCheckboxColumn,
-    showRowCheckbox,
   } = useStandardTableConfig();
 
   const selectedIds = useStandardTableState().selectedIds.selectedIds;
   const tableId = useStandardTableId();
   const onKeyDown = useOnKeyDownContext();
   const { numNavigableColumns } = useColumnIndexPerColumnIdContext();
+  const stickyPropsPerColumnContext = useStickyPropsPerColumnContext();
 
   const isSelected = useMemo(() => {
     const itemId = keyResolver(item);
@@ -55,7 +55,6 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
   const {
     itemValueResolver,
     itemLabelFormatter,
-    flex = 1,
     width,
     minWidth,
     justifyContentCell = "flex-start",
@@ -65,9 +64,7 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
     isEditable,
     onChange,
     disableGridCell,
-    sticky,
     zIndex,
-    left,
   } = useColumnConfigById(columnId);
 
   const itemValue = useMemo(() => {
@@ -141,6 +138,8 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
     borderLeft
   );
 
+  const stickyProps = stickyPropsPerColumnContext[columnId];
+
   return (
     <StandardTableCellUi
       enableGridCell={enableGridCell && !disableGridCell}
@@ -150,18 +149,28 @@ export const StandardTableCell = React.memo(function StandardTableCell<TItem>({
       minWidth={minWidth}
       justifyContent={justifyContentCell}
       borderLeft={activeBorderLeft}
-      flex={flex}
       background={currentBackground}
-      sticky={sticky}
-      zIndex={zIndex}
-      left={
-        sticky && showRowCheckbox && stickyCheckboxColumn && left == null
-          ? "45px"
-          : sticky && !stickyCheckboxColumn && left == null
-          ? "0px"
-          : left
+      sticky={stickyProps.sticky}
+      zIndex={
+        zIndex ?? stickyProps.sticky
+          ? "var(--swui-sticky-column-z-index)"
+          : undefined
       }
-      shadow={sticky ? "var(--swui-sticky-column-shadow-right)" : undefined}
+      left={stickyProps.left}
+      right={stickyProps.right}
+      shadow={
+        stickyProps.sticky &&
+        stickyProps.type === "last-group" &&
+        stickyProps.isFirstColumnInLastGroup
+          ? "var(--swui-sticky-column-shadow-left)"
+          : stickyProps.sticky &&
+            stickyProps.type === "column" &&
+            stickyProps.right
+          ? "var(--swui-sticky-column-shadow-left)"
+          : stickyProps.sticky
+          ? "var(--swui-sticky-column-shadow-right)"
+          : undefined
+      }
       onKeyDown={onKeyDownHandler}
     >
       {content}