+++ packages/grid/src/features/standard-table/stories/ExpandableRows.stories.tsx
@@ -36,6 +36,37 @@ export const ExpandableRows = () => {
   return <StandardTable items={items} config={config} />;
 };
 
+export const WithSticky = () => {
+  const { items, onChangeNumPassengers } = useListState(mockedItems);
+
+  const config: StandardTableConfig<ListItem, keyof ListItem> = {
+    enableExpandCollapse: true,
+    stickyHeader: true,
+    stickyCheckboxColumn: true,
+    renderRowExpansion: (item) => (
+      <Box spacing indent>
+        <Text>Name: {item.name}</Text>
+      </Box>
+    ),
+    ...standardTableConfigForStories,
+    columns: {
+      ...standardTableConfigForStories.columns,
+      numPassengers: {
+        ...standardTableConfigForStories.columns.numPassengers,
+        onChange: onChangeNumPassengers,
+      },
+    },
+  };
+
+  return (
+    <div style={{ width: "700px", height: "450px", overflow: "scroll" }}>
+      <div style={{ width: "800px", height: "550px" }}>
+        <StandardTable items={items} config={config} />
+      </div>
+    </div>
+  );
+};
+
 export const SomeExpandableRows = () => {
   const { items, onChangeNumPassengers } = useListState(mockedItems);